adds WebAPIMessageBridge and WebAPIPresenceBridge classes that intercept OSCAR buddy/ICBM events and push them to WebAPI event queues, while also handling AMF3 encoding fixes for Gromit compatibility including timestamp conversions and proper event data structuring.

+++ server/webapi/adapters.go
@@ -141,7 +141,7 @@ func ICBMToWebAPIEvent(icbm wire.SNAC_0x04_0x07_ICBMChannelMsgToClient) (state.W
 		Data: state.IMEvent{
 			From:      senderScreenName,
 			Message:   messageText,
-			Timestamp: time.Now().Unix(),
+			Timestamp: float64(time.Now().Unix()),
 			AutoResp:  autoResponse,
 		},
 	}
